pyproject.toml
[tool.poetry]
name = "weatherapi"
version = "0.1.0"
description = ""
authors = ["william.cuesta <william09y@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "8.0.33"
aiohttp = "3.8.4"
fastapi = "0.98.0"
apscheduler = "3.10.1"
uvicorn = "0.22.0"
pytest = "7.4.0"
flake8 = "6.0.0"
httpx = "0.24.1"
fastapi-cache2 = "0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
